Move SystemTransaction into its own module
