+++ src/bin/bench-tps.rs
@@ -19,9 +19,10 @@ use solana::metrics;
 use solana::ncp::Ncp;
 use solana::service::Service;
 use solana::signature::{read_keypair, GenKeys, Keypair, KeypairUtil};
+use solana::system_transaction::SystemTransaction;
 use solana::thin_client::{poll_gossip_for_leader, ThinClient};
 use solana::timing::{duration_as_ms, duration_as_s};
-use solana::transaction::{SystemTransaction, Transaction};
+use solana::transaction::Transaction;
 use solana::wallet::request_airdrop;
 use solana::window::default_window;
 use std::collections::VecDeque;
