+++ src/lib.rs
@@ -58,6 +58,7 @@ pub mod storage_program;
 pub mod store_ledger_stage;
 pub mod streamer;
 pub mod system_program;
+pub mod system_transaction;
 pub mod thin_client;
 pub mod tictactoe_program;
 pub mod timing;
